[Frontend] WordCloud widget refetch (#14448)
