+++ opencti-platform/opencti-front/src/components/dashboard/WidgetWordCloud.tsx
@@ -1,9 +1,9 @@
 import React, { useMemo } from 'react';
 import { useTheme } from '@mui/styles';
 import ReactWordcloud, { Props } from 'react-wordcloud';
-// eslint-disable-next-line import/no-extraneous-dependencies
+
 import 'tippy.js/dist/tippy.css';
-// eslint-disable-next-line import/no-extraneous-dependencies
+
 import 'tippy.js/animations/scale.css';
 import type { Theme } from '../Theme';
 import { colors } from '../../utils/Charts';
@@ -11,7 +11,7 @@ import useDistributionGraphData from '../../utils/hooks/useDistributionGraphData
 
 interface WidgetWordCloudProps {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  data: any[];
+  data: readonly any[];
   groupBy: string;
 }
